Adds Support for Publishing Native Terms. (#3)
* Adds Support for Publishing Native Terms.

New callback = cerlnan_avro:publish which publishes single records, or
batches, as object container files.  Users must provide:

* The type name for the records in the batch.

* The erlavro compatible schema for the records in the batch.

* A single proplist or list of proplists containing key/values
appropriate for the given type/schema.

* Adds publish to ExCernan.Avro.

In addition to the ability to publish terms from Elixir, we also
introduce cerlnan_avro_schema which serves as an abstraction layer,
preventing erlavro from leaking out into user code.

* Cleanup Types + Fix (Most) Dialyzer Errors

* Make Dialyzer Run Cleanly.

We now:

* try/catch around erlavro serialization.  Erlavro can raise a number of
exceptions and this was throwing dialyzer due to erlavro's type specs.
We now return a generic error tuple and print the original stacktrace on
erlavro exception.

* Exempt unit tests from dialyzer by only including them in modules with
TEST is defined as a macro.

* Byte buffer no longer blocks indefinitely waiting for buffer contents.
A default timeout of 1 second is provided, returning an error tuple when
a timeout occurs.