Rename Project to cerlnan_avro. (#9)
* Switch Cerlnan to Focus on Erlang Only

The intended use case of using a mix backed umbrella app as a
dependency in other apps was not supported by mix. This changeset
switches the preferred build system to rebar3 with extensions
for users to opt-out of certain subapps at build time.

* Remove Elixir from cerlnan_avro's README

* Rename Project cerlnan_avro

After some (re?) education on mix and rebar3 dependencies, umbrella
projects in general are not supported as dependencies by modern build
tools.  As such, this project is being renamed cerlnan_avro such that
other projects can make use of it while a cerlnan is reworked.